refactor: ParkingOperation 등록 API에서 요청 vo, dto명을 서비스 메서드명과 일치
